+++ torch/csrc/api/include/torch/nn/modules.h
@@ -4,7 +4,7 @@
 #include <torch/nn/modules/batchnorm.h>
 #include <torch/nn/modules/conv.h>
 #include <torch/nn/modules/dropout.h>
-#include <torch/nn/modules/embedding.h>
+#include <torch/nn/modules/sparse.h>
 #include <torch/nn/modules/fold.h>
 #include <torch/nn/modules/functional.h>
 #include <torch/nn/modules/linear.h>
