+++ torch/csrc/api/src/nn/modules/sparse.cpp
@@ -1,4 +1,4 @@
-#include <torch/nn/modules/embedding.h>
+#include <torch/nn/modules/sparse.h>
 
 #include <torch/types.h>
 #include <torch/utils.h>
